+++ R/autoplot-combined-dist.R
@@ -210,6 +210,8 @@ tidy_combined_autoplot <- function(.data, .plot_type = "density", .line_size = .
     }
 
     if (.geom_smooth) {
+        max_dy <- max(data_tbl$dy)
+
         plt <- plt +
             ggplot2::geom_smooth(
                 ggplot2::aes(
@@ -218,7 +220,8 @@ tidy_combined_autoplot <- function(.data, .plot_type = "density", .line_size = .
                 se = FALSE,
                 color = "black",
                 linetype = "dashed"
-            )
+            ) +
+            ggplot2::xlim(0, max_dy)
     }
 
     if (.geom_jitter) {