+++ NEWS.md
@@ -9,7 +9,8 @@ to say `scale` instead of `rate` in the returned estimated parameters.
 None
 
 ## Minor Fixes and Improvements
-None
+1. Fix #90 - Make sure when `.geom_smooth` is set to TRUE that `ggplot2::xlim(0, max_dy)`
+is set.
 
 # TidyDensity 1.0.0